add NotFoundError to framework
